// build/typechain/contracts/Staking.d.ts
export class Staking extends Contract {
// ...
  filters: {
    AllocationClosed(
      indexer: string | null,
      subgraphDeploymentID: BytesLike | null,
      epoch: null,
      tokens: null,
      allocationID: string | null,
      effectiveAllocation: null,
      sender: null,
      poi: null,
      isDelegator: null
// ...
  };
// ...
}